Refactor share handlers into reusable hook
