+++ components/panels/ChatPane.tsx
@@ -15,7 +15,6 @@ import { getRandomWelcome } from '@/lib/welcomeMessages';
 import { useActiveContext } from '@/lib/context';
 import { isFollowUp } from '@/lib/followup';
 import { detectFollowupIntent } from '@/lib/intents';
-import { BRAND_NAME } from "@/lib/brand";
 import SuggestionChips from "@/components/chat/SuggestionChips";
 import SuggestBar from "@/components/suggest/SuggestBar";
 import ComposerFocus from "@/components/chat/ComposerFocus";
@@ -49,9 +48,7 @@ import { getUserPosition, fetchNearby, geocodeArea, type NearbyKind, type Nearby
 import { formatTrialBriefMarkdown } from "@/lib/trials/brief";
 import { useIsAiDocMode } from "@/hooks/useIsAiDocMode";
 import { pushToast } from "@/lib/ui/toast";
-import { trackEvent } from "@/lib/analytics/events";
-import { exportMessageCardToPng } from "@/lib/share/exportImage";
-import { buildShareCaption, buildShareIntentCaption } from "@/lib/share/caption";
+import { useShareActions } from "@/lib/share/useShareActions";
 
 const AIDOC_UI = process.env.NEXT_PUBLIC_AIDOC_UI === '1';
 const AIDOC_PREFLIGHT = process.env.NEXT_PUBLIC_AIDOC_PREFLIGHT === '1';
@@ -91,18 +88,6 @@ type RefreshSnapshot = {
   context?: string | null;
 };
 
-type ShareTargetState = {
-  conversationId: string;
-  messageId: string;
-  domId: string;
-  plainText: string;
-  mdText?: string | null;
-  mode: 'patient' | 'doctor' | 'research' | 'therapy';
-  research: boolean;
-};
-
-const SHARE_FALLBACK_MESSAGE = 'Shared response unavailable.';
-
 const cloneFilters = (value: any) => {
   if (value == null) return null;
   try {
@@ -124,24 +109,6 @@ const getMessageTextFromDom = (domId: string, fallback: string) => {
   const content = node.querySelector('.message-content');
   const text = content ? (content as HTMLElement).innerText : node.innerText;
   return (text || fallback || '').trim();
-};
-
-const fallbackCopyToClipboard = (text: string) => {
-  const textarea = document.createElement('textarea');
-  textarea.value = text;
-  textarea.setAttribute('readonly', 'true');
-  textarea.style.position = 'fixed';
-  textarea.style.opacity = '0';
-  textarea.style.pointerEvents = 'none';
-  textarea.style.left = '-9999px';
-  document.body.appendChild(textarea);
-  textarea.focus();
-  textarea.select();
-  const successful = document.execCommand('copy');
-  document.body.removeChild(textarea);
-  if (!successful) {
-    throw new Error('copy_failed');
-  }
 };
 
 const formatTrendDate = (iso?: string) => {
@@ -734,17 +701,6 @@ export default function ChatPane({ inputRef: externalInputRef }: { inputRef?: Re
   const [loadingAction, setLoadingAction] = useState<null | 'simpler' | 'doctor' | 'next'>(null);
   const [labSummary, setLabSummary] = useState<any | null>(null);
   const [refreshingMessages, setRefreshingMessages] = useState<Record<string, boolean>>({});
-  const [shareTarget, setShareTarget] = useState<ShareTargetState | null>(null);
-  const [shareBusy, setShareBusy] = useState<
-    | null
-    | 'link'
-    | 'download'
-    | 'caption'
-    | 'system'
-    | 'x'
-    | 'facebook'
-  >(null);
-  const [shareLinks, setShareLinks] = useState<Record<string, string>>({});
   const abortRef = useRef<AbortController | null>(null);
   const chatRef = useRef<HTMLDivElement>(null);
   const inputRef =
@@ -752,13 +708,22 @@ export default function ChatPane({ inputRef: externalInputRef }: { inputRef?: Re
     (useRef<HTMLTextAreaElement>(null) as RefObject<HTMLTextAreaElement>);
   const previewUrlsRef = useRef<string[]>([]);
   const queueAbortRef = useRef<AbortController | null>(null);
-  const shareLinksRef = useRef<Record<string, string>>({});
-  const shareInFlightRef = useRef<Map<string, Promise<string>>>(new Map());
   const { filters } = useResearchFilters();
 
-  useEffect(() => {
-    shareLinksRef.current = shareLinks;
-  }, [shareLinks]);
+  const {
+    shareTarget,
+    shareBusy,
+    sharePreview,
+    systemShareSupported,
+    openShare,
+    closeShare,
+    handleCopyShareLink,
+    handleDownloadShareImage,
+    handleShareX,
+    handleShareFacebook,
+    handleSystemShare,
+    handleCopyShareCaption,
+  } = useShareActions();
 
   const sp = useSearchParams();
   const isAiDocMode = useIsAiDocMode();
@@ -3087,284 +3052,6 @@ ${systemCommon}` + baseSys;
     }
   }, [messages, conversationId, setFromChat, setUi]);
 
-  const systemShareSupported =
-    typeof navigator !== 'undefined' && 'share' in navigator && typeof (navigator as any).share === 'function';
-
-  const ensureShareLink = useCallback(
-    async (target: ShareTargetState) => {
-      const cached = shareLinksRef.current[target.messageId];
-      if (cached) return cached;
-
-      const inflight = shareInFlightRef.current.get(target.messageId);
-      if (inflight) return inflight;
-
-      const request = (async () => {
-        const fallbackMessage = SHARE_FALLBACK_MESSAGE;
-        const safePlainText = (target.plainText || '').trim();
-        const safeMdText = (target.mdText || '').trim();
-        const plainTextPayload = safePlainText || safeMdText || fallbackMessage;
-        const mdTextPayload = safeMdText || undefined;
-
-        const res = await fetch('/api/share', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({
-            conversationId: target.conversationId,
-            messageId: target.messageId,
-            plainText: plainTextPayload,
-            mdText: mdTextPayload,
-            mode: target.mode,
-            research: target.research,
-          }),
-        });
-        if (!res.ok) throw new Error('share_failed');
-        const data = await res.json().catch(() => null);
-        const absoluteUrlRaw = typeof data?.absoluteUrl === 'string' ? data.absoluteUrl.trim() : '';
-        const slug = typeof data?.slug === 'string' ? data.slug : '';
-
-        const normalizedAppUrl = (() => {
-          const envUrl = (process.env.NEXT_PUBLIC_APP_URL ?? '').trim();
-          if (envUrl) return envUrl.replace(/\/$/, '');
-          if (typeof window !== 'undefined' && window.location?.origin) {
-            return window.location.origin.replace(/\/$/, '');
-          }
-          return '';
-        })();
-
-        const resolveAbsoluteUrl = (value: string) => {
-          if (!value) return '';
-          if (/^https?:\/\//i.test(value)) return value;
-          if (value.startsWith('//')) {
-            return `https:${value}`;
-          }
-          if (value.startsWith('/')) {
-            return normalizedAppUrl ? `${normalizedAppUrl}${value}` : '';
-          }
-          return normalizedAppUrl ? `${normalizedAppUrl}/${value.replace(/^\/+/, '')}` : '';
-        };
-
-        const absoluteUrl = resolveAbsoluteUrl(absoluteUrlRaw);
-
-        let shareUrl = '';
-        if (absoluteUrl) {
-          shareUrl = absoluteUrl;
-        } else if (slug) {
-          if (!normalizedAppUrl) throw new Error('share_failed');
-          shareUrl = `${normalizedAppUrl}/s/${slug}`;
-        } else {
-          throw new Error('share_failed');
-        }
-
-        shareLinksRef.current[target.messageId] = shareUrl;
-        setShareLinks((prev) => {
-          if (prev[target.messageId]) return prev;
-          return { ...prev, [target.messageId]: shareUrl };
-        });
-        return shareUrl;
-      })();
-      shareInFlightRef.current.set(target.messageId, request);
-      try {
-        return await request;
-      } finally {
-        shareInFlightRef.current.delete(target.messageId);
-      }
-    },
-    [setShareLinks]
-  );
-
-  const closeShare = useCallback(() => {
-    setShareTarget(null);
-    setShareBusy(null);
-  }, []);
-
-  const openShare = useCallback((target: ShareTargetState) => {
-    const fallbackMessage = SHARE_FALLBACK_MESSAGE;
-    const normalizedPlain = target.plainText.trim();
-    const normalizedMd = target.mdText?.trim() ?? '';
-    const normalizedTarget: ShareTargetState = {
-      ...target,
-      plainText: normalizedPlain || normalizedMd || fallbackMessage,
-      mdText: normalizedMd || undefined,
-    };
-    setShareTarget(normalizedTarget);
-    setShareBusy(null);
-    trackEvent('share_opened', {
-      conversationId: normalizedTarget.conversationId,
-      messageId: normalizedTarget.messageId,
-      mode: normalizedTarget.mode,
-      research: normalizedTarget.research,
-    });
-  }, []);
-
-  const handleCopyShareLink = useCallback(async () => {
-    if (!shareTarget) return;
-    setShareBusy('link');
-    try {
-      const url = await ensureShareLink(shareTarget);
-      try {
-        if (!navigator?.clipboard?.writeText) throw new Error('clipboard_unsupported');
-        await navigator.clipboard.writeText(url);
-      } catch {
-        fallbackCopyToClipboard(url);
-      }
-      pushToast({ title: 'Link copied' });
-      trackEvent('share_copy_link', {
-        conversationId: shareTarget.conversationId,
-        messageId: shareTarget.messageId,
-        mode: shareTarget.mode,
-        research: shareTarget.research,
-      });
-    } catch {
-      pushToast({ title: 'Could not copy link', variant: 'destructive' });
-    } finally {
-      setShareBusy(null);
-    }
-  }, [shareTarget, ensureShareLink]);
-
-  const handleDownloadShareImage = useCallback(async () => {
-    if (!shareTarget) return;
-    setShareBusy('download');
-    try {
-      const node = document.getElementById(shareTarget.domId);
-      if (!node) throw new Error('missing_node');
-      const modeLabel =
-        shareTarget.mode === 'doctor'
-          ? 'Doctor mode'
-          : shareTarget.mode === 'research'
-          ? 'Research mode'
-          : shareTarget.mode === 'therapy'
-          ? 'Therapy mode'
-          : 'Patient mode';
-      const dataUrl = await exportMessageCardToPng(node, {
-        brand: BRAND_NAME,
-        modeLabel,
-        timestamp: new Date(),
-        fallbackText: shareTarget.plainText,
-      });
-      const anchor = document.createElement('a');
-      anchor.href = dataUrl;
-      anchor.download = 'medx-answer.png';
-      document.body.appendChild(anchor);
-      anchor.click();
-      document.body.removeChild(anchor);
-      pushToast({ title: 'Image saved' });
-      trackEvent('share_download_image', {
-        conversationId: shareTarget.conversationId,
-        messageId: shareTarget.messageId,
-        mode: shareTarget.mode,
-        research: shareTarget.research,
-      });
-    } catch {
-      pushToast({ title: 'Could not save image', variant: 'destructive' });
-    } finally {
-      setShareBusy(null);
-    }
-  }, [shareTarget]);
-
-  const handleShareX = useCallback(async () => {
-    if (!shareTarget) return;
-    const win = window.open('about:blank', '_blank', 'noopener,noreferrer');
-    setShareBusy('x');
-    try {
-      const url = await ensureShareLink(shareTarget);
-      const caption = buildShareIntentCaption(shareTarget.plainText, BRAND_NAME);
-      const intentUrl = `https://twitter.com/intent/tweet?text=${encodeURIComponent(caption)}&url=${encodeURIComponent(url)}`;
-      if (win) {
-        win.location.href = intentUrl;
-      } else {
-        window.open(intentUrl, '_blank', 'noopener,noreferrer');
-      }
-      trackEvent('share_x', {
-        conversationId: shareTarget.conversationId,
-        messageId: shareTarget.messageId,
-        mode: shareTarget.mode,
-        research: shareTarget.research,
-      });
-    } catch {
-      if (win) {
-        win.close();
-      }
-      pushToast({ title: 'Could not open X share', variant: 'destructive' });
-    } finally {
-      setShareBusy(null);
-    }
-  }, [shareTarget, ensureShareLink]);
-
-  const handleShareFacebook = useCallback(async () => {
-    if (!shareTarget) return;
-    const win = window.open('about:blank', '_blank', 'noopener,noreferrer');
-    setShareBusy('facebook');
-    try {
-      const url = await ensureShareLink(shareTarget);
-      const intentUrl = `https://www.facebook.com/sharer/sharer.php?u=${encodeURIComponent(url)}`;
-      if (win) {
-        win.location.href = intentUrl;
-      } else {
-        window.open(intentUrl, '_blank', 'noopener,noreferrer');
-      }
-      trackEvent('share_facebook', {
-        conversationId: shareTarget.conversationId,
-        messageId: shareTarget.messageId,
-        mode: shareTarget.mode,
-        research: shareTarget.research,
-      });
-    } catch {
-      if (win) {
-        win.close();
-      }
-      pushToast({ title: 'Could not open Facebook share', variant: 'destructive' });
-    } finally {
-      setShareBusy(null);
-    }
-  }, [shareTarget, ensureShareLink]);
-
-  const handleSystemShare = useCallback(async () => {
-    if (!shareTarget) return;
-    if (!systemShareSupported || !navigator?.share) {
-      pushToast({ title: 'Device sharing not available', variant: 'destructive' });
-      return;
-    }
-    setShareBusy('system');
-    try {
-      const url = await ensureShareLink(shareTarget);
-      const caption = buildShareCaption(shareTarget.plainText, BRAND_NAME, url);
-      await navigator.share({ title: 'MedX Answer', text: caption, url });
-      trackEvent('share_system', {
-        conversationId: shareTarget.conversationId,
-        messageId: shareTarget.messageId,
-        mode: shareTarget.mode,
-        research: shareTarget.research,
-      });
-    } catch (err: any) {
-      if (err?.name !== 'AbortError') {
-        pushToast({ title: 'Share cancelled', variant: 'destructive' });
-      }
-    } finally {
-      setShareBusy(null);
-    }
-  }, [shareTarget, ensureShareLink, systemShareSupported]);
-
-  const handleCopyShareCaption = useCallback(async () => {
-    if (!shareTarget) return;
-    setShareBusy('caption');
-    try {
-      const url = await ensureShareLink(shareTarget);
-      const caption = buildShareCaption(shareTarget.plainText, BRAND_NAME, url);
-      try {
-        if (!navigator?.clipboard?.writeText) throw new Error('clipboard_unsupported');
-        await navigator.clipboard.writeText(caption);
-      } catch {
-        fallbackCopyToClipboard(caption);
-      }
-      pushToast({ title: 'Caption copied' });
-    } catch {
-      pushToast({ title: 'Could not copy caption', variant: 'destructive' });
-    } finally {
-      setShareBusy(null);
-    }
-  }, [shareTarget, ensureShareLink]);
-
-  const sharePreview = shareTarget?.plainText ?? '';
 
   const renderedMessages = useMemo(
     () =>
@@ -3494,13 +3181,13 @@ ${systemCommon}` + baseSys;
                 conversationId,
                 messageId: m.id!,
                 domId,
-              plainText: plain,
-              mdText: messageContent,
-              mode: currentMode,
-              research: researchMode,
-            });
-          }
-        : undefined;
+                plainText: plain,
+                mdText: messageContent,
+                mode: currentMode,
+                research: researchMode,
+              });
+            }
+          : undefined;
 
         return (
           <div key={derivedKey} className="space-y-2">
